Centralise a load of exception definitions
David MacIver started moving us in this direction, but we never followed
up on it properly. Having the exceptions in a standalone module like
this leads to much more sane import graphs and avoids the potential for
circular import hell which is where we were heading.
